fix : profileUser 엔티티 삭제
- user 엔티티 사용
